forge-llm-service/requirements.txt
flask>=2.0.1
openai>=1.0.0
python-dotenv>=0.19.0